Added some tests and changed PromotionConfig. Changed PromotionConfig freeAmount to Integer and added tests for MORE_OFF promotion type.

+++ src/main/java/dev/aziz/grocerystore/entities/PromotionConfig.java
@@ -33,7 +33,7 @@ public class PromotionConfig {
     private Integer minimumAmount;
 
     @NotNull(message = "Free amount should not be empty")
-    private Double freeAmount;
+    private Integer freeAmount;
 
     @OneToOne
     private Item item;
